chore: [APP-30025] [APP-30026] [APP-30027] [APP-30028] resolve git comments
