+++ lib/enums.ts
@@ -1,7 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export
-export enum EVENT_TYPES {
+export enum EventTypes {
   PAGE,
-  KEY_ARROW_LEFT,
-  KEY_ARROW_RIGHT,
   CUSTOM,
 }
